Remove channel runtime process-local state
