+++ agents/Aevatar.GAgents.ChannelRuntime/LarkConversationAdapterFactory.cs
@@ -1,92 +1,39 @@
-using System.Collections.Concurrent;
 using System.Text.Json;
 using Aevatar.Foundation.Abstractions.Credentials;
 using Aevatar.GAgents.Channel.Abstractions;
 using Aevatar.GAgents.Channel.Lark;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
 namespace Aevatar.GAgents.ChannelRuntime;
 
-internal sealed class LarkConversationAdapterRegistry : IAsyncDisposable
+internal sealed class LarkConversationAdapterFactory
 {
     private readonly LarkMessageComposer _composer;
     private readonly LarkPayloadRedactor _payloadRedactor;
     private readonly IHttpClientFactory _httpClientFactory;
     private readonly ILoggerFactory _loggerFactory;
-    private readonly IServiceProvider _services;
-    private readonly ConcurrentDictionary<string, AdapterRegistration> _registrations = new(StringComparer.Ordinal);
+    private readonly ICredentialProvider? _credentialProvider;
 
-    public LarkConversationAdapterRegistry(
+    public LarkConversationAdapterFactory(
         LarkMessageComposer composer,
         LarkPayloadRedactor payloadRedactor,
         IHttpClientFactory httpClientFactory,
         ILoggerFactory loggerFactory,
-        IServiceProvider services)
+        ICredentialProvider? credentialProvider = null)
     {
         _composer = composer ?? throw new ArgumentNullException(nameof(composer));
         _payloadRedactor = payloadRedactor ?? throw new ArgumentNullException(nameof(payloadRedactor));
         _httpClientFactory = httpClientFactory ?? throw new ArgumentNullException(nameof(httpClientFactory));
         _loggerFactory = loggerFactory ?? throw new ArgumentNullException(nameof(loggerFactory));
-        _services = services ?? throw new ArgumentNullException(nameof(services));
+        _credentialProvider = credentialProvider;
     }
 
-    public async Task<LarkChannelAdapter> GetAsync(ChannelBotRegistrationEntry registration, CancellationToken ct)
+    public async Task<LarkChannelAdapter> CreateAsync(ChannelBotRegistrationEntry registration, CancellationToken ct)
     {
         ArgumentNullException.ThrowIfNull(registration);
         ct.ThrowIfCancellationRequested();
 
-        var snapshot = await AdapterSnapshot.FromAsync(registration, _services.GetService<ICredentialProvider>(), ct);
-        if (_registrations.TryGetValue(snapshot.RegistrationId, out var existing) &&
-            existing.Snapshot == snapshot)
-        {
-            return existing.Adapter;
-        }
-
-        var adapter = await CreateAsync(snapshot, ct);
-        var replacement = new AdapterRegistration(snapshot, adapter);
-        while (true)
-        {
-            ct.ThrowIfCancellationRequested();
-
-            if (_registrations.TryAdd(snapshot.RegistrationId, replacement))
-                return adapter;
-
-            if (!_registrations.TryGetValue(snapshot.RegistrationId, out var observed))
-                continue;
-
-            if (observed.Snapshot == snapshot)
-            {
-                await adapter.StopReceivingAsync(CancellationToken.None);
-                return observed.Adapter;
-            }
-
-            if (_registrations.TryUpdate(snapshot.RegistrationId, replacement, observed))
-            {
-                await observed.Adapter.StopReceivingAsync(CancellationToken.None);
-                return adapter;
-            }
-        }
-    }
-
-    public async ValueTask DisposeAsync()
-    {
-        foreach (var registration in _registrations.Values)
-        {
-            try
-            {
-                await registration.Adapter.StopReceivingAsync(CancellationToken.None);
-            }
-            catch
-            {
-            }
-        }
-
-        _registrations.Clear();
-    }
-
-    private async Task<LarkChannelAdapter> CreateAsync(AdapterSnapshot snapshot, CancellationToken ct)
-    {
+        var snapshot = await AdapterSnapshot.FromAsync(registration, _credentialProvider, ct);
         var adapter = new LarkChannelAdapter(
             new StaticCredentialProvider(snapshot.SecretJson),
             _composer,
@@ -109,16 +56,11 @@ internal sealed class LarkConversationAdapterRegistry : IAsyncDisposable
         return adapter;
     }
 
-    private sealed record AdapterRegistration(
-        AdapterSnapshot Snapshot,
-        LarkChannelAdapter Adapter);
-
     private sealed record AdapterSnapshot(
         string RegistrationId,
         string ScopeId,
         string VerificationToken,
         string BindingCredentialRef,
-        string CredentialRef,
         string SecretJson)
     {
         public static async Task<AdapterSnapshot> FromAsync(
@@ -144,7 +86,6 @@ internal sealed class LarkConversationAdapterRegistry : IAsyncDisposable
                 registration.ScopeId ?? string.Empty,
                 registration.GetVerificationToken(),
                 bindingCredentialRef,
-                credentialRef,
                 secretJson);
         }
