+++ agents/Aevatar.GAgents.ChannelRuntime/LarkConversationIngressRuntime.cs
@@ -13,18 +13,18 @@ internal interface ILarkConversationIngressRuntime
 
 internal sealed class LarkConversationIngressRuntime : ILarkConversationIngressRuntime
 {
-    private readonly LarkConversationAdapterRegistry _adapterRegistry;
+    private readonly LarkConversationAdapterFactory _adapterFactory;
     private readonly ILarkConversationInbox _inbox;
     private readonly ILogger<LarkConversationIngressRuntime> _logger;
     private readonly IChannelRuntimeDiagnostics? _diagnostics;
 
     public LarkConversationIngressRuntime(
-        LarkConversationAdapterRegistry adapterRegistry,
+        LarkConversationAdapterFactory adapterFactory,
         ILarkConversationInbox inbox,
         ILogger<LarkConversationIngressRuntime> logger,
         IChannelRuntimeDiagnostics? diagnostics = null)
     {
-        _adapterRegistry = adapterRegistry ?? throw new ArgumentNullException(nameof(adapterRegistry));
+        _adapterFactory = adapterFactory ?? throw new ArgumentNullException(nameof(adapterFactory));
         _inbox = inbox ?? throw new ArgumentNullException(nameof(inbox));
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         _diagnostics = diagnostics;
@@ -37,35 +37,42 @@ internal sealed class LarkConversationIngressRuntime : ILarkConversationIngressR
 
         try
         {
-            var adapter = await _adapterRegistry.GetAsync(registration, ct);
-            var request = await BuildWebhookRequestAsync(http, ct);
-            var response = await adapter.HandleWebhookAsync(request, ct);
-
-            if (!string.IsNullOrWhiteSpace(response.ResponseBody))
+            var adapter = await _adapterFactory.CreateAsync(registration, ct);
+            try
             {
-                RecordDiagnostic("Callback:verified", registration.Platform, registration.Id, "lark_webhook_verification");
-                return Results.Content(
-                    response.ResponseBody,
-                    "application/json",
-                    Encoding.UTF8,
-                    response.StatusCode);
-            }
+                var request = await BuildWebhookRequestAsync(http, ct);
+                var response = await adapter.HandleWebhookAsync(request, ct);
 
-            if (response.StatusCode != StatusCodes.Status200OK)
+                if (!string.IsNullOrWhiteSpace(response.ResponseBody))
+                {
+                    RecordDiagnostic("Callback:verified", registration.Platform, registration.Id, "lark_webhook_verification");
+                    return Results.Content(
+                        response.ResponseBody,
+                        "application/json",
+                        Encoding.UTF8,
+                        response.StatusCode);
+                }
+
+                if (response.StatusCode != StatusCodes.Status200OK)
+                {
+                    RecordDiagnostic("Callback:error", registration.Platform, registration.Id, $"webhook_rejected:{response.StatusCode}");
+                    return Results.StatusCode(response.StatusCode);
+                }
+
+                if (response.Activity is null)
+                {
+                    RecordDiagnostic("Callback:ignored", registration.Platform, registration.Id, "adapter_returned_null");
+                    return Results.Ok(new { status = "ignored" });
+                }
+
+                await _inbox.EnqueueAsync(response.Activity, ct);
+                RecordDiagnostic("Callback:accepted", registration.Platform, registration.Id, "durable_inbox_committed");
+                return Results.Ok(new { status = "accepted" });
+            }
+            finally
             {
-                RecordDiagnostic("Callback:error", registration.Platform, registration.Id, $"webhook_rejected:{response.StatusCode}");
-                return Results.StatusCode(response.StatusCode);
+                await adapter.StopReceivingAsync(CancellationToken.None);
             }
-
-            if (response.Activity is null)
-            {
-                RecordDiagnostic("Callback:ignored", registration.Platform, registration.Id, "adapter_returned_null");
-                return Results.Ok(new { status = "ignored" });
-            }
-
-            await _inbox.EnqueueAsync(response.Activity, ct);
-            RecordDiagnostic("Callback:accepted", registration.Platform, registration.Id, "durable_inbox_committed");
-            return Results.Ok(new { status = "accepted" });
         }
         catch (Exception ex)
         {
